requirements.txt
dj-rest-auth==4.0.1
Django==4.1.5
django-agora==0.4
django-allauth==0.54.0
django-compat==1.0.15
django-cors-headers==3.14.0
django-environ==0.10.0
django-extensions==3.2.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==10.0.0
